// frontend/src/api/generated/serializers/Auth.ts
// Typelizer digest adf04512c0a04c0ecb181c643ea0eed3
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {User, Organization, Membership} from "./index"

type Auth = {
  user: User;
  impersonator: User | null;
  organization: Organization;
  membership: Membership;
  organizations: Array<Organization>;
  superadmin: boolean;
  onboardingRequired: boolean;
  sessionId: string;
  sudoUntil: string | null;
}

export type { Auth };