// frontend/src/api/generated/serializers/AuthSession.ts
// Typelizer digest a3f52b342a828398be5bc8772b754d7c
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {User} from "./index"

type AuthSession = {
  token: string | null;
  newAccount: boolean;
  expiresAt: string;
  sudoUntil: string | null;
  user: User;
  impersonator: User | null;
  canManage: boolean;
}

export type { AuthSession };